backend/requirements.txt
# Core dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy[asyncio]==2.0.23
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8

# AI/ML
openai==1.10.0
httpx==0.26.0

# Async tasks
celery==5.3.4
redis==5.0.1

# Image processing
Pillow==10.1.0
requests==2.31.0

# Utilities
pydantic==2.5.0
pydantic-settings==2.1.0
python-dateutil==2.8.2
aiofiles==23.2.1

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.1
isort==5.13.2
flake8==6.1.0

# Monitoring
sentry-sdk[fastapi]==1.40.0
prometheus-fastapi-instrumentator==6.0.0

# WebSocket support
websockets==12.0
python-socketio==5.10.0
